refactor(hammerspoon): init.lua を config/*.lua へ責務分割し薄いローダー化
一枚岩だった init.lua を責務ごとに config/*.lua へ分割し、init.lua は
require で読み込むだけの薄いローダーにした。挙動は不変（純リファクタ）。

- config/reload.lua      設定オートリロード（pathwatcher + notify）
- config/behavior.lua    Preferences > Behavior（hs.autoLaunch 等6設定）
- config/app_toggle.lua  アプリ トグル（ホットキー）
- config/ime.lua         En/Ja 切替（eikana は GC 防止でグローバル維持）
- config/caffeine.lua    Caffeine（スリープ防止）

追加/削除は config/ にファイルを置き init.lua の require 一覧を増減するだけ。
Hammerspoon は既定で ~/.hammerspoon を package.path に含むため
require("config.x") で config/x.lua を解決でき、pathwatcher の再帰監視により
config/*.lua の編集でも自動リロードされる。install.sh は cp -fr で config/ ごと配布。

挙動不変の確認: 旧 init.lua と新 config 群で「実行 Lua 行の集合」が完全一致（各93行、
コメント/空行除外・sort 済み diff が空）。中身の変更（Spoon 管理の SpoonInstall 化・
WezTerm→Ghostty 掃除）は後続 PR で行う。
