+++ roles/hammerspoon/.hammerspoon/config/behavior.lua
@@ -1,0 +1,7 @@
+-- Preferences > Behavior（GUI のチェック状態をコードで固定・冪等）
+hs.autoLaunch(true)                     -- Launch Hammerspoon at login: ON
+hs.automaticallyCheckForUpdates(true)   -- Check for updates: ON
+hs.dockIcon(false)                      -- Show dock icon: OFF
+hs.menuIcon(true)                       -- Show menu icon: ON
+hs.consoleOnTop(false)                  -- Keep Console window on top: OFF
+hs.uploadCrashData(false)               -- Send crash data: OFF
