+++ roles/hammerspoon/.hammerspoon/config/caffeine.lua
@@ -1,0 +1,7 @@
+-- Caffeine（スリープ防止）: 起動時にオン・hyper+F6 でトグル
+
+-- Load Spoons: https://github.com/Hammerspoon/Spoons
+hs.loadSpoon("Caffeine")
+spoon.Caffeine:bindHotkeys({toggle={hyper, "f6"}})
+spoon.Caffeine:start()
+spoon.Caffeine:clicked()
