+++ roles/hammerspoon/.hammerspoon/config/reload.lua
@@ -1,0 +1,16 @@
+-- 設定オートリロード（~/.hammerspoon 配下の .lua 保存で hs.reload）
+
+-- Config loaded auto, when save init.lua file
+function reloadConfig(files)
+    doReload = false
+    for _,file in pairs(files) do
+        if file:sub(-4) == ".lua" then
+            doReload = true
+        end
+    end
+    if doReload then
+        hs.reload()
+    end
+end
+hs.pathwatcher.new(os.getenv("HOME") .. "/.hammerspoon/", reloadConfig):start()
+hs.notify.new({title="Hammerspoon", informativeText="Config loaded"}):send()
